+++ tests/__init__.py
@@ -14,4 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License
 """
 
-REPORT_PORTAL_SERVICE = 'robotframework_reportportal.service.RPClient'
+REPORT_PORTAL_SERVICE = 'reportportal_client.RPClient'