+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 ### Added
+- `RP_CLIENT_TYPE` configuration variable, by @HardNorth
+- `RP_CONNECT_TIMEOUT` and `RP_READ_TIMEOUT` configuration variables, by @HardNorth
+### Changed
+- Client version updated on [5.5.0](https://github.com/reportportal/client-Python/releases/tag/5.5.0), by @HardNorth
+### Removed
+- Dependency on `six`, by @HardNorth
+
+## [5.4.0]
+### Added
 - `RP_LAUNCH_UUID_PRINT` and `RP_LAUNCH_UUID_PRINT_OUTPUT` configuration variables, by @HardNorth
 ### Changed
 - Internal item list was replaced with LifoQueue, by @HardNorth